Making class file to new classmates
